casino check for duplicates

+++ app/routes/bets.js
@@ -124,12 +124,13 @@ const updateParentUserBalanceTemp = async (parentUsersIds, matchId = 0, bet, run
      // user.availableBalance = UseravailableBalancePrev-finalShareAmountInLoss;
      
      user.availableBalance = prevBalance +   (-finalShareAmountInLoss)
-
+     if(bet.fancyData == null && userId==45341){
      await expPositive.deleteMany({
       userId:user.userId,
       roundId: bet.marketId,
       userFrom:bet.userId
     });
+  }
 
      await expPositive.create({
       userId:user.userId,
@@ -153,11 +154,13 @@ const updateParentUserBalanceTemp = async (parentUsersIds, matchId = 0, bet, run
       user.exposure = userPrevExposure-finalShareAmountInLoss;
       user.availableBalance = prevBalance +   (userPrevExposure-finalShareAmountInLoss)
    // user.availableBalance = UseravailableBalancePrev - finalShareAmountInLoss;
+   if(bet.fancyData == null && userId==45341){
    await expPositive.deleteMany({
     userId:user.userId,
     roundId: bet.marketId,
     userFrom:bet.userId
   });
+}
         await expPositive.create({
           userId:user.userId,
           userFrom:bet.userId,
@@ -227,11 +230,13 @@ const updateParentUserBalanceTemp = async (parentUsersIds, matchId = 0, bet, run
     user.exposure = -finalShareAmountInLoss;
     //user.availableBalance = UseravailableBalancePrev-finalShareAmountInLoss;
     user.availableBalance = prevBalance + (-finalShareAmountInLoss);
+    if(bet.fancyData == null && userId==45341){
     await expPositive.deleteMany({
       userId:user.userId,
       roundId: bet.marketId,
       userFrom:bet.userId
     });
+  }
     await expPositive.create({
       userId:user.userId,
       userFrom:bet.userId,
@@ -265,12 +270,13 @@ const updateParentUserBalanceTemp = async (parentUsersIds, matchId = 0, bet, run
     user.exposure = -finalShareAmountInLoss;
     //user.availableBalance = prevAdjustedAvailableBalance-finalShareAmountInLoss;
     user.availableBalance = prevBalance + (-finalShareAmountInLoss);
-    
+    if(bet.fancyData == null && userId==45341){
     await expPositive.deleteMany({
       userId:user.userId,
       roundId: bet.marketId,
       userFrom:bet.userId
     });
+  }
     await expPositive.create({
       userId:user.userId,
       userFrom:bet.userId,
@@ -300,12 +306,13 @@ const updateParentUserBalanceTemp = async (parentUsersIds, matchId = 0, bet, run
     user.exposure = prevAdjustedExposure - finalShareAmountInLoss;
     //user.availableBalance =prevAdjustedAvailableBalance - finalShareAmountInLoss;
     user.availableBalance =prevBalance + (prevAdjustedExposure - finalShareAmountInLoss);
-
+    if(bet.fancyData == null && userId==45341){
     await expPositive.deleteMany({
       userId:user.userId,
       roundId: bet.marketId,
       userFrom:bet.userId
     });
+  }
     await expPositive.create({
       userId:user.userId,
       userFrom:bet.userId,
@@ -3781,11 +3788,12 @@ const placeBet = async (req, res) => {
           const betId = mongoose.Types.ObjectId(result._id); // Convert if necessary
           
           const bet = await Bets.findOne({ _id: betId });
-
+          if(bet.fancyData == null && userId==45341){
            await expPositive.deleteMany({
             userId: userId,
             roundId: bet.marketId
           });
+          }
 
          await  expPositive.create({
             userId:userId,
